+++ app/src/main/java/com/tolstykh/eatABurrita/location/LocationModule.kt
@@ -1,6 +1,7 @@
 package com.tolstykh.eatABurrita.location
 
 import android.content.Context
+import com.google.android.gms.location.GeofencingClient
 import com.google.android.gms.location.LocationServices
 import dagger.Module
 import dagger.Provides
@@ -20,4 +21,10 @@ object LocationModule {
         context,
         LocationServices.getFusedLocationProviderClient(context)
     )
+
+    @Singleton
+    @Provides
+    fun provideGeofencingClient(
+        @ApplicationContext context: Context
+    ): GeofencingClient = LocationServices.getGeofencingClient(context)
 }